Chore(tests, BDD): rename integration-test to tests dir
Rename integration-test dir to tests following go idiomatic
package naming convention.
