+++ buildscripts/test-cov.sh
@@ -3,7 +3,7 @@
 set -e
 echo "" > coverage.txt
 
-for d in $(go list ./... | grep -v 'vendor\|pkg/apis\|pkg/client/generated\|integration-tests'); do
+for d in $(go list ./... | grep -v 'vendor\|pkg/apis\|pkg/client/generated\|tests'); do
     #TODO - Include -race while creating the coverage profile.
     go test -coverprofile=profile.out -covermode=atomic $d
     if [ -f profile.out ]; then
@@ -13,7 +13,7 @@ for d in $(go list ./... | grep -v 'vendor\|pkg/apis\|pkg/client/generated\|inte
 done
 
 echo "Running go vet on integration-test"
-for d in $(go list ./... | grep 'integration-tests'); do
+for d in $(go list ./... | grep 'tests'); do
     #TODO - Currently we are checking only on integration-test.
     go vet $d
 done
