+++ tests/admission/admission_test.go
@@ -19,11 +19,11 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/openebs/maya/integration-tests/artifacts"
-	framework "github.com/openebs/maya/integration-tests/framework/v1alpha1"
 	k8s "github.com/openebs/maya/pkg/client/k8s/v1alpha1"
 	cv "github.com/openebs/maya/pkg/cstorvolume/v1alpha1"
 	pvc "github.com/openebs/maya/pkg/kubernetes/persistentvolumeclaim/v1alpha1"
+	"github.com/openebs/maya/tests/artifacts"
+	framework "github.com/openebs/maya/tests/framework/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
