+++ tests/artifacts/cstor-pvc.yaml
