+++ tests/artifacts/cvr-schema.yaml
