+++ tests/artifacts/sc-hostdevice.yaml
