+++ tests/artifacts/sc-hostpath.yaml
