+++ tests/artifacts/storageclass-single-replica.yaml
