+++ tests/jiva/node-stickiness/replicadeletion/jiva_replica_deletion_test.go
@@ -19,12 +19,12 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/openebs/maya/integration-tests/artifacts"
-	nodestickiness "github.com/openebs/maya/integration-tests/jiva/node-stickiness"
 	pvc "github.com/openebs/maya/pkg/kubernetes/persistentvolumeclaim/v1alpha1"
 	pod "github.com/openebs/maya/pkg/kubernetes/pod/v1alpha1"
 	sc "github.com/openebs/maya/pkg/kubernetes/storageclass/v1alpha1"
 	unstruct "github.com/openebs/maya/pkg/unstruct/v1alpha2"
+	"github.com/openebs/maya/tests/artifacts"
+	nodestickiness "github.com/openebs/maya/tests/jiva/node-stickiness"
 	corev1 "k8s.io/api/core/v1"
 	k8serror "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
