+++ tests/sts/sts_suite_test.go
@@ -26,9 +26,9 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/openebs/maya/integration-tests/artifacts"
-	"github.com/openebs/maya/integration-tests/kubernetes"
 	pod "github.com/openebs/maya/pkg/kubernetes/pod/v1alpha1"
+	"github.com/openebs/maya/tests/artifacts"
+	"github.com/openebs/maya/tests/kubernetes"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	// auth plugins
